feat(lobby): M3 T2——大厅协议/登录接入网关/JSON 配置 + T1 昵称唯一
- lobby.go：大厅协议 200-231 编解码（登录注册/好友/房间/背包 + roundtrip 测试）
- uid 统一 uint32（账号 uid 直接进房——对局身份一致）；Token 改 4+8+HMAC
- 网关登录态：会话挂账号 uid/昵称——未登录只放行注册/登录（其余拦截）
- msgid.go 清理（M2 双进程消息废弃——登录移入大厅段 200）
- config：JSON 配置（example 上传 + config.json gitignore 不追踪）+ 启动日志
- main 组装：config → MySQL/Mem → lobby → 网关（内存兜底）
- 昵称唯一（users.nickname UNIQUE + SearchByNickname——好友按昵称搜）
- 测试：注册登录流/未登录拒绝/昵称重复（gateway/lobby/protocol 全绿 -race）

+++ db/schema.sql
@@ -4,7 +4,7 @@ USE lurkspire;
 
 -- 用户：账号密码登录 + 昵称显示（独立且唯一——好友按昵称搜索添加）；装备持久化
 CREATE TABLE IF NOT EXISTS users (
-    id            BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+    id            INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     account       VARCHAR(64)  NOT NULL UNIQUE,
     password_hash VARCHAR(100) NOT NULL,          -- bcrypt
     nickname      VARCHAR(32)  NOT NULL UNIQUE,   -- 昵称唯一（好友搜索凭据）
@@ -14,9 +14,9 @@ CREATE TABLE IF NOT EXISTS users (
 
 -- 好友关系（双向：A-B 一条关系存两行，friend_a < friend_b 恒成立 + 各自方向视图查询）
 CREATE TABLE IF NOT EXISTS friends (
-    id         BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-    user_a     BIGINT UNSIGNED NOT NULL,           -- 小 id
-    user_b     BIGINT UNSIGNED NOT NULL,           -- 大 id
+    id         INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+    user_a     INT UNSIGNED NOT NULL,           -- 小 id
+    user_b     INT UNSIGNED NOT NULL,           -- 大 id
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     UNIQUE KEY uk_pair (user_a, user_b)
 ) ENGINE=InnoDB;
